SENT: Lua glue: getSentValue returns ETB position now
Can be tested with following LUA script:
setTickRate(100)
function onTick()
	pos = getSentValue(1)
	setLuaGauge(1, 10000 * pos)
	print(pos)
end

+++ firmware/hw_layer/drivers/sent/sent.h
@@ -19,4 +19,4 @@ void SENT_ISR_Handler(uint8_t ch, uint16_t val_res);
 void startSent(void);
 void stopSent(void);
 
-float getSentValue(int index);
+float getSentValue(size_t index);
